ci/cd: remove daily workflow trigger
FYI @AlexAxthelm I think it's unnecessary to trigger new scenario builds every day. 

It ends up just making it very difficult to validate if we are choosing the correct version of scenario outputs to include in a PROD run of data prep. 

Suggest leaving it to only merges to `main` and `workflow_dispatch` triggers.